Reorganized FASM generation scripts
